+++ skiko/import-generator/src/jvmMain/kotlin/org/jetbrains/skiko/ImportGeneratorExtension.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.skia
+package org.jetbrains.skiko
 
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
