+++ skiko/import-generator/src/jvmMain/kotlin/org/jetbrains/skiko/options.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.skia
+package org.jetbrains.skiko
 
 const val PATH_OPTION_NAME = "import-generator-path"
 const val PREFIX_OPTION_NAME = "import-generator-prefix"
