+++ skiko/import-generator/src/jvmMain/resources/META-INF/services/org.jetbrains.kotlin.compiler.plugin.CommandLineProcessor
@@ -1,1 +1,1 @@
-org.jetbrains.skia.ImportGeneratorCommandLineProcessor
+org.jetbrains.skiko.ImportGeneratorCommandLineProcessor
