+++ skiko/import-generator/src/jvmMain/resources/META-INF/services/org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar
@@ -1,1 +1,1 @@
-org.jetbrains.skia.ImportGeneratorRegistrar
+org.jetbrains.skiko.ImportGeneratorRegistrar
